Add Handlebars Helper
We let it slip that people could write reviews with markdown, but haven’t yet implemented that functionality. We also realized that we’ve been slightly overcharging people. Turns out that the price of our products is actually in cents! So 999 cents should really be $9.99. Follow the tasks below to create some Handlebars Helpers to fix these problems.

Also, we’ve added two additional libraries for you to use in this challenge. They’re unrelated to Ember, but they’ll help out!

- Accounting.js - a tiny JavaScript library for number, money and currency formatting.
- Markdown.js - A JavaScript markdown parser.